Add React Router And Refactor Some Components
