+++ package.json
@@ -18,6 +18,7 @@
     "@storybook/addons": "^5.3.17",
     "@storybook/react": "^5.3.17",
     "@types/node": "^13.9.2",
+    "@types/reach__router": "^1.3.1",
     "@types/react": "^16.9.23",
     "@types/react-dom": "^16.9.5",
     "babel-core": "^6.26.3",
