+++ packages/app/src/components/App.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Explorer from "./pages/Explorer/Explorer";
+import Explorer from "../pages/Explorer/Explorer";
 
 export default function App() {
   return <Explorer />;
