+++ packages/components/src/index.ts
@@ -1,5 +1,4 @@
 export { default as Block } from "./components/Block";
-export { default as BlockDetails } from "./components/BlockDetails";
 export { default as Card } from "./components/Card";
 export { default as CardStack } from "./components/CardStack";
 export { default as ContentLabel } from "./components/ContentLabel";
@@ -9,5 +8,4 @@ export { default as GridItem } from "./components/GridItem";
 export { default as Label } from "./components/Label";
 export { default as Link } from "./components/Link";
 export { default as SearchBar } from "./components/SearchBar";
-export { default as TabLink } from "./components/TabLink";
 export { default as TabNavigation } from "./components/TabNavigation";
